T now compactable with json

+++ src/ThreeHiggs/TransitionFinder.py
@@ -171,7 +171,7 @@ def traceFreeEnergyMinimum(effectivePotential,
     for T in TRange:
         if verbose:
             print (f'Start of temp = {T} loop')
-        minimizationResults["T"].append(T)
+        minimizationResults["T"].append(float(T))
         goalRGScale =  T ## Final scale in 3D -check if goalRGscale is ever different from just T
 
         matchingScale = 4.0*np.pi*math.exp(-EulerGamma) * T ## Scale that minimises T dependent logs
@@ -220,7 +220,7 @@ def traceFreeEnergyMinimum(effectivePotential,
             if effectivePotential.bReachedUltraSoftScale(minimumLocation, ## Check if ultra soft scale reached
                                                          T, 
                                                          verbose = verbose): 
-                minimizationResults["UltraSoftTemp"] = int(T) ## If reached then set that as the ultra soft temp
+                minimizationResults["UltraSoftTemp"] = float(T) ## If reached then set that as the ultra soft temp
                                                          ##- this will stop the first if statement from passing
         
         if minimizationResults["bIsPerturbative"]: ##If the potential was perturbative check if it still is
